+++ noodles-bam/examples/bam_query.rs
@@ -9,7 +9,7 @@ use std::{env, io, path::PathBuf};
 use noodles_bam as bam;
 use noodles_sam::{
     self as sam,
-    alignment::{io::Writer, RecordBuf},
+    alignment::{io::Write, RecordBuf},
 };
 
 const UNMAPPED: &str = "*";